% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.SPI.R
\name{MR.SPI}
\alias{MR.SPI}
\title{Mendelian randomization that first selects valid IVs and then perform post-selection inference}
\usage{
MR.SPI(
  gamma,
  Gamma,
  se_gamma,
  se_Gamma,
  n1,
  n2,
  freq = NULL,
  tuning = 1,
  max_clique = TRUE,
  robust = FALSE,
  alpha = 0.05,
  a = 0.6,
  Sampling = TRUE,
  rho = NULL,
  M = 1000,
  prop = 0.1,
  filtering = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gamma}{a numeric vector of GWAS summary statistics of the exposure}

\item{Gamma}{a numeric vector of GWAS summary statistics of the outcome}

\item{se_gamma}{a numeric vector of standard errors of gamma}

\item{se_Gamma}{a numeric vector of standard errors of Gamma}

\item{n1}{the sample size of GWAS summary statistics  of the exposure}

\item{n2}{the sample size of GWAS summary statistics  of the outcome}

\item{freq}{a numeric vector of allele frequencies to adjust for the regression coefficients and the standard errors, with default NULL}

\item{tuning}{a numeric scalar value tuning parameter for selecting the valid IVs, with default 1}

\item{max_clique}{an option to replace the majority and plurality voting procedures with finding maximal clique in the IV voting matrix, with default TRUE}

\item{robust}{an option to produce robust confidence interval, with default FALSE}

\item{alpha}{a numeric scalar value between 0 and 1 indicating the significance level for the confidence interval, with default 0.05}

\item{a}{grid size for constructing beta grids, with default 0.6}

\item{Sampling}{use the sampling method if TRUE; else use the proposed searching method, with default TRUE (only applicable if unif=TRUE)}

\item{rho}{a numeric scalar denoting thresholding level used in the sampling property (only applicable if unif=TRUE)}

\item{M}{sampling times, with default value 1000 (only applicable if unif=TRUE)}

\item{prop}{proportion of intervals kept when sampling, with default value 0.1 (only applicable if unif=TRUE)}

\item{filtering}{Filtering the resampled data or not, with default value TRUE (only applicable if unif=TRUE)}

\item{verbose}{a logical value to determine whether to print detailed results, the default value is TRUE.}
}
\value{
\code{MR.SPI} return a list containing the following:
    \item{\code{betaHat}}{a numeric scalar denoting the estimate of causal effect.}
    \item{\code{beta.sdHat}}{a numeric scalar denoting the estimated standard error of betaHat.}
    \item{\code{ci}}{a two dimensional numeric vector denoting the 1-alpha confidence intervals for betaHat with lower and upper endpoints.}
    \item{\code{SHat}}{a numeric vector denoting the set of relevant IVs}
    \item{\code{VHat}}{a numeric vector denoting the set of valid IVs}
    \item{\code{voting.mat}}{a matrix of voting. if (i,j)-th entry=1, then i-th SNP and j-th SNP vote for each other to be a valid IV.}
}
\description{
This package is used for conduting two-sample Mendelian randomization with indepnendent SNPs by selecting valid IVs that satisfy the core assumptions and inferring the causal effect.
}
